+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Column order and coloring in volume list
+- Show layer-list instead of ports column in resource-list
 
 ## [1.23.2] - 2024-09-25